Embed metal view in engine
