--- NeonLightsTestApp/ViewModels/NeonView.swift
@@ -1,47 +1,0 @@
-//
-//  NeonView.swift
-//  NeonLightsTestApp
-//
-//  Created by Aleksandr Tiulpanov on 08/09/2025.
-//
-
-
-//
-//  NeonView.swift
-//  NeonLightsTestApp
-//
-//  Created by Aleksandr Tiulpanov on 08/09/2025.
-//
-
-import SwiftUI
-import MetalKit
-import NeonEngine
-
-public struct NeonView: UIViewRepresentable {
-    let renderer: NeonRenderer
-
-    public func makeUIView(context: Context) -> MTKView {
-        // Create the MTKView using the same device as the renderer to ensure
-        // all Metal resources are compatible.
-        print("🪟 Creating MTKView")
-        let view = MTKView(frame: .zero, device: renderer.device)
-        renderer.configure(view: view)
-        return view
-    }
-
-    public func updateUIView(_ uiView: MTKView, context: Context) {
-        // Keep the drawable size in sync with SwiftUI layout changes so Metal
-        // renders at the correct resolution and aspect ratio.
-        let scale = uiView.contentScaleFactor
-        let size = CGSize(width: uiView.bounds.width * scale,
-                          height: uiView.bounds.height * scale)
-        uiView.drawableSize = size
-        // Explicitly notify the renderer about the new drawable size so it can
-        // allocate or resize its internal buffers immediately. Relying on
-        // MTKView's delegate callback can be racy when the size changes during
-        // SwiftUI layout updates, leading to a black screen if the renderer
-        // hasn't yet provisioned its textures.
-        renderer.mtkView(uiView, drawableSizeWillChange: size)
-        print("📐 updateUIView -> drawableSize: \(size)")
-    }
-}
